requirements.txt
flask
flask-bootstrap
flask-bcrypt
flask-login
flask-mail
flask-sqlalchemy
flask-wtf
flask
PyJWT
wtforms